+++ CHANGELOG.md
@@ -9,7 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
-- `timer.rs` refactoring
+- `timer.rs` refactoring [#553]
+- Add `rcc::Instance` trait [#557]
+- Demacrofy `RxDma`/`TxDma` implementations [#554]
+
+[#553]: https://github.com/stm32-rs/stm32f1xx-hal/pull/553
+[#554]: https://github.com/stm32-rs/stm32f1xx-hal/pull/554
+[#557]: https://github.com/stm32-rs/stm32f1xx-hal/pull/557
 
 ### Added